forms and add script, venue class own file
