+++ .github/ISSUE_TEMPLATE/add_venue.md
@@ -1,0 +1,32 @@
+---
+name: Add new venue
+about: Update list of local venues
+title: 'Add Venue'
+labels: 'add venue'
+assignees: '@MatBenfield'
+
+please fill in the following values:
+
+<!-- name starts-->
+name
+<!-- name ends-->
+
+<!-- desc starts-->
+description
+<!-- desc ends-->
+
+<!-- website starts-->
+website url
+<!-- website ends-->
+
+<!-- facebook starts-->
+facebook url
+<!-- facebook ends-->
+
+<!-- twitter starts-->
+twitter handle
+<!-- twitter ends-->
+
+<!-- instagram starts-->
+instagram url
+<!-- instagram ends-->
