+++ .github/ISSUE_TEMPLATE/hijack_request.md
@@ -1,0 +1,13 @@
+---
+name: Hijack Request
+about: Template to request hijacking of the randomiser for 1 week
+title: 'Hijack [venue name]'
+labels: 'Hijack'
+assignees: '@MatBenfield'
+
+---
+
+- [ ] Ensure the venue name is in the title
+- [ ] provide any social media accounts
+- [ ] provide a description of the venue, feel free to include any timely offers.
+- [ ] Arrange for payment via paypal
